[Personal] CAES-1149: Fix key generating
